Update new logic authentication

+++ controllers/postController.js
@@ -1,18 +1,19 @@
 const Post = require("../models/Post");
+const User = require("../models/User");
+const jwt = require("jsonwebtoken");
 
 // Get All Posts
 exports.getAllPosts = async (req, res, next) => {
   try {
     const posts = await Post.find({})
       .populate("author", "name")
-      .select("content createdAt"); //_id được tự động thêm vào. Nếu muốn loại bỏ _id thì sẽ thêm vào select('-_id')
+      .select("content createdAt");
     res.status(200).json({
       status: "success",
       results: posts.length,
       data: { posts },
     });
   } catch (error) {
-    // res.json(error);
     next(error);
   }
 };
@@ -21,14 +22,12 @@ exports.getAllPosts = async (req, res, next) => {
 exports.createOnePost = async (req, res, next) => {
   try {
     const { userId } = req.user;
-    console.log(userId);
     const post = await Post.create({ ...req.body, author: userId });
     res.status(200).json({
       status: "success",
       data: { post },
     });
   } catch (error) {
-    // res.json(error);
     next(error);
   }
 };
@@ -37,17 +36,27 @@ exports.createOnePost = async (req, res, next) => {
 exports.updateOnePost = async (req, res, next) => {
   try {
     const { postId } = req.params;
-    const post = await Post.findByIdAndUpdate(
+    const post = await Post.findById(postId);
+
+    // Check if the user is the author of the post
+    if (post.author.toString() !== req.user.userId) {
+      const error = new Error("You do not have permission to update this post");
+      error.statusCode = 403;
+      throw error;
+    }
+
+    // Update the post
+    const updatedPost = await Post.findByIdAndUpdate(
       postId,
       { ...req.body },
-      { new: true, runValidator: true }
+      { new: true, runValidators: true }
     );
+
     res.status(200).json({
       status: "success",
-      data: { post },
+      data: { post: updatedPost },
     });
   } catch (error) {
-    // res.json(error);
     next(error);
   }
 };
@@ -56,13 +65,23 @@ exports.updateOnePost = async (req, res, next) => {
 exports.deleteOnePost = async (req, res, next) => {
   try {
     const { postId } = req.params;
+    const post = await Post.findById(postId);
+
+    // Check if the user is the author of the post
+    if (post.author.toString() !== req.user.userId) {
+      const error = new Error("You do not have permission to delete this post");
+      error.statusCode = 403;
+      throw error;
+    }
+
+    // Delete the post
     await Post.findByIdAndDelete(postId);
+
     res.status(200).json({
       status: "success",
       message: "Post has been deleted",
     });
   } catch (error) {
-    // res.json(error);
     next(error);
   }
 };
